Feat: Modified add-task-UI
feat: added task-page-out
chore: changed structure of add-task
